refactor: Add the data.py and move get_coefficient function.
