+++ src/sum_dirac_dfcoef/sum_dirac_dfcoef.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 
 import argparse
-import bisect
 import copy
 from io import TextIOWrapper
 import os
@@ -9,9 +8,10 @@ import sys
 
 
 from .args import args
-from .coefficient import Coefficient, Coefficients
+from .coefficient import get_coefficient
+from .data import Data_MO, Data_All_MO
 from .utils import debug_print, space_separated_parsing
-from .functions_info import AtomInfo, FunctionsInfo, get_functions_info
+from .functions_info import AtomInfo, get_functions_info
 
 
 def is_this_row_for_coefficients(words: "list[str]") -> bool:
@@ -56,84 +56,6 @@ def get_dirac_filename(args: "argparse.Namespace") -> str:
     return args.file
 
 
-def get_coefficient(line: str, orbitals: FunctionsInfo, idx: int) -> Coefficient:
-    """
-    Nested functions to get coefficient
-    (e.g.)
-    sym_and_atom_and_orb_str = "B3gCldyz"
-    symmetry_type = "B3g"
-    atom_type = "Cl"
-    orbital_type = "dyz"
-    """
-    # ref (print ): https://gitlab.com/dirac/dirac/-/blob/b10f505a6f00c29a062f5cad70ca156e72e012d7/src/dirac/dirout.F#L388-389
-    # ref (format): https://gitlab.com/dirac/dirac/-/blob/b10f505a6f00c29a062f5cad70ca156e72e012d7/src/dirac/dirout.F#L453
-    # FORMAT(3X,I5,2X,A12,2X,4F14.10)
-    # https://gitlab.com/dirac/dirac/-/blob/b10f505a6f00c29a062f5cad70ca156e72e012d7/src/dirac/dirtra.F#L168-169
-
-    def is_float(parameter: str):
-        if not parameter.isdecimal():
-            try:
-                float(parameter)
-                return True
-            except ValueError:
-                return False
-        else:
-            return False
-
-    def parse_line(line: "str") -> Coefficient:
-        """
-        This function parses the line that contains the coefficient.
-
-        line write source: https://gitlab.com/dirac/dirac/-/blob/b10f505a6f00c29a062f5cad70ca156e72e012d7/src/dirac/dirout.F#L440-442
-        line format source: https://gitlab.com/dirac/dirac/-/blob/b10f505a6f00c29a062f5cad70ca156e72e012d7/src/dirac/dirout.F#L453
-        line format : FORMAT(3X,I5,2X,A12,2X,4F14.10)
-        """
-        words = line.split()
-        # JS (I5): Serial number of the vector
-        vec_num = int(words[0])
-
-        #
-        # PLABEL(IPLAB(IBAS(IFRP)+JS,2),2) (A12): Information about the vector to identify the vector
-        #
-        # PLABEL source: https://gitlab.com/dirac/dirac/-/blob/b10f505a6f00c29a062f5cad70ca156e72e012d7/src/dirac/dirtra.F#L168-169
-        # PLABEL(NLAB,2) = CLS(IC)//' '//REP(IRP)//NAMN(ICENT)(1:3)//GTOTYP(ITYP)
-        # CLS (A1): https://gitlab.com/dirac/dirac/-/blob/b10f505a6f00c29a062f5cad70ca156e72e012d7/src/dirac/dirtra.F#L45
-        # REP (A3): https://gitlab.com/dirac/dirac/-/blob/b10f505a6f00c29a062f5cad70ca156e72e012d7/src/include/pgroup.h#L16
-        # NAMN (A3, defined as A4, but only (1:3) is used): https://gitlab.com/dirac/dirac/-/blob/b10f505a6f00c29a062f5cad70ca156e72e012d7/src/include/nuclei.h#L25
-        # GTOTYP (A4): https://gitlab.com/dirac/dirac/-/blob/b10f505a6f00c29a062f5cad70ca156e72e012d7/src/include/ccom.h#L8
-        component_func = "large orbitals" if line[10] == "L" else ("small orbitals" if line[10] == "S" else "")  # CLS
-        symmetry_label = line[12:15].strip()  # REP (e.g. "Ag ")
-        atom_label = line[15:18].strip()  # NAMN (e.g. "Cm "), atom_labe="Cm"
-        function_label = line[12:22].strip().replace(" ", "")  # REP + NAMN + GTOTYP (e.g. "Ag Cm s   " => "AgCms")
-
-        # COEF (4F14.10)
-        # coefficients = [line[24:38], line[38:52], line[52:66], line[66:80]]
-        coef_num = 4
-        coef_len = 14
-        coef_start_idx = 24
-        coefficient = sum(
-            [
-                pow(float(line[i : i + coef_len]), 2) if is_float(line[i : i + coef_len]) else pow(-100, 2)
-                for i in range(coef_start_idx, coef_start_idx + coef_len * coef_num, coef_len)
-            ]
-        )
-
-        need_identifier = True if len(orbitals[component_func][symmetry_label][atom_label]) > 1 or orbitals[component_func][symmetry_label][atom_label][idx].mul > 1 else False
-        key_idx = orbitals[component_func][symmetry_label][atom_label][idx].start_idx
-        multiplication = int(orbitals[component_func][symmetry_label][atom_label][key_idx].mul)
-
-        return Coefficient(
-            vector_num=vec_num, function_label=function_label, need_identifier=need_identifier, coefficient=coefficient, start_idx=key_idx, multiplication=multiplication
-        )
-
-    """
-    Main function to get coefficient
-    """
-    coef = parse_line(line)
-
-    return coef
-
-
 def check_start_vector_print(words: "list[str]") -> bool:
     # ****************************** Vector print ******************************
     if len(words) < 4:
@@ -166,34 +88,6 @@ def get_output_path(args: "argparse.Namespace") -> str:
     return output_path
 
 
-def create_results_for_current_mo(args: "argparse.Namespace", coefficients: Coefficients) -> Coefficients:
-    """
-    Create results for current MO
-    """
-
-    def coef_sort_and_remove(coefficients: Coefficients) -> Coefficients:
-        """
-        Sort coefficients and remove elements below threshold
-        """
-
-        def remove_elements_below_threshold() -> Coefficients:
-            """
-            Remove elements below threshold
-            """
-
-            coefficients.coef_list.sort(key=lambda x: x.coefficient)  # ascending order
-            idx = bisect.bisect_left([c.coefficient / coefficients.norm_const_sum * 100 for c in coefficients.coef_list], args.threshold)
-            coefficients.coef_list = coefficients.coef_list[idx:] if idx < len(coefficients.coef_list) else []
-            coefficients.coef_list.sort(key=lambda x: (-x.coefficient, x.vector_num))  # descending order
-            return coefficients
-
-        coefficients.coef_list = [v for v in coefficients.coef_dict.values()]
-        coefficients = remove_elements_below_threshold()
-        return coefficients
-
-    return copy.deepcopy(coef_sort_and_remove(coefficients))
-
-
 def should_write_positronic_results_to_file(args: "argparse.Namespace") -> bool:
     if args.all_write or args.positronic_write:
         return True
@@ -208,7 +102,7 @@ def should_write_electronic_results_to_file(args: "argparse.Namespace") -> bool:
         return False
 
 
-def write_results(args: "argparse.Namespace", file: TextIOWrapper, data_all_mo: "list[Coefficients]") -> None:
+def write_results(args: "argparse.Namespace", file: TextIOWrapper, data_all_mo: "list[Data_MO]") -> None:
     """
     Write results to stdout
     """
@@ -239,18 +133,15 @@ def main() -> None:
     is_reading_coefficients: bool = False
     start_vector_print: bool = False
     is_electronic: bool = False
-    electron_number: int = 0
-    prev_electron_number: int = electron_number
     mo_energy: float = 0.0
     mo_sym_type: str = ""
     start_idx: int = 1
-    coefficients: Coefficients = Coefficients()
 
     dirac_filename: str = get_dirac_filename(args)
     dirac_output = open(dirac_filename, encoding="utf-8")
     orbitals = get_functions_info(dirac_output)
-    data_all_electronic_mo: "list[Coefficients]" = []
-    data_all_positronic_mo: "list[Coefficients]" = []
+    data_mo = Data_MO()
+    data_all_mo = Data_All_MO()
     used_atom_info: dict[str, AtomInfo] = dict()
     current_atom_info = AtomInfo()
     with open(dirac_filename, encoding="utf-8") as f:
@@ -268,12 +159,12 @@ def main() -> None:
             elif need_to_skip_this_line(words):
                 if need_to_create_results_for_current_mo(words, is_reading_coefficients):
                     start_mo_coefficients = False
-                    data = create_results_for_current_mo(args, coefficients)
+                    data_mo.fileter_coefficients_by_threshold()
                     if is_electronic:
-                        data_all_electronic_mo.append(data)
+                        data_all_mo.electronic.append(copy.deepcopy(data_mo))
                     else:  # Positronic
-                        data_all_positronic_mo.append(data)
-                    debug_print(f"End of reading {electron_number}th MO")
+                        data_all_mo.positronic.append(copy.deepcopy(data_mo))
+                    debug_print(f"End of reading {data_mo.electron_num}th MO")
                     is_reading_coefficients = False
 
             elif need_to_start_mo_section(words, start_mo_coefficients):
@@ -291,22 +182,23 @@ def main() -> None:
                 else:
                     raise Exception("Unknown MO type")
                 try:
-                    electron_number = int(words[-2][:-1].replace("no.", ""))
+                    electron_num = int(words[-2][:-1].replace("no.", ""))
                 except ValueError:
                     # If *** is printed, we have no information about what number this MO is.
-                    # Therefore, we assume that electron_number is the next number after prev_electron_number.
-                    electron_number = prev_electron_number + 1
-                prev_electron_number = electron_number
+                    # Therefore, we assume that electron_num is the next number after prev_electron_num.
+                    prev_electron_num = data_mo.electron_num  # prev_electron is the number of electrons of the previous MO
+                    electron_num = prev_electron_num + 1
                 mo_energy = float(words[-1])
                 mo_info = (
-                    f"{mo_sym_type} {electron_number}"
+                    f"{mo_sym_type} {electron_num}"
                     if args.compress
-                    else (f"Electronic no. {electron_number} {mo_sym_type}" if is_electronic else f"Positronic no. {electron_number} {mo_sym_type}")
+                    else (f"Electronic no. {electron_num} {mo_sym_type}" if is_electronic else f"Positronic no. {electron_num} {mo_sym_type}")
                 )
                 # Here is the start point of reading coefficients of the current MO
-                coefficients.reset()  # reset coefficients because we need to delete coefficients of the previous MO
-                coefficients.mo_energy = mo_energy
-                coefficients.mo_info = mo_info
+                data_mo.reset()  # reset data_mo because we need to delete data_mo of the previous MO
+                data_mo.electron_num = electron_num
+                data_mo.mo_energy = mo_energy
+                data_mo.mo_info = mo_info
                 used_atom_info.clear()  # reset used_atom_info because we need to delete used_atom_info of the previous MO
 
             elif check_end_vector_print(
@@ -346,21 +238,19 @@ def main() -> None:
                     used_atom_info[function_without_gto_type] = copy.deepcopy(current_atom_info)
 
                 current_atom_info.decrement_function(gto_type)
-                coef = get_coefficient(line, orbitals, start_idx)
-                coefficients.add_coefficient(coef)
+                data_mo.add_coefficient(get_coefficient(line, orbitals, start_idx))
 
     # End of reading file
     # Write results to the file
     file = open(get_output_path(args), "w")
+    if not args.no_sort:
+        data_all_mo.electronic.sort(key=lambda x: x.mo_energy)
+        data_all_mo.positronic.sort(key=lambda x: x.mo_energy)
     if should_write_positronic_results_to_file(args):  # Positronic
         # Write positronic results to the file
-        if not args.no_sort:
-            data_all_positronic_mo.sort(key=lambda x: x.mo_energy)
-        write_results(args, file, data_all_positronic_mo)
+        write_results(args, file, data_all_mo.positronic)
         file.write("\n")  # Add a blank line
     if should_write_electronic_results_to_file(args):  # Electronic
         # Write electronic results to the file
-        if not args.no_sort:
-            data_all_electronic_mo.sort(key=lambda x: x.mo_energy)
-        write_results(args, file, data_all_electronic_mo)
+        write_results(args, file, data_all_mo.electronic)
     file.close()
